chore: add db module and colocate json-ld types. Also rdf types in corresponding module
