+++ web-registry/src/components/molecules/BatchMetadata/BatchMetadata.tsx
@@ -2,7 +2,8 @@ import { Box } from '@mui/material';
 
 import { Flex } from 'web-components/lib/components/box';
 import { Title } from 'web-components/lib/components/typography';
-import { CreditBatchMetadataIntersectionLD } from 'web-components/lib/types/rdf/credit-batch-intersection-ld';
+
+import { CreditBatchMetadataIntersectionLD } from 'lib/db/types/json-ld';
 
 import { BatchMetadataAdditionalInfo } from './BatchMetadata.AdditionalInfo';
