+++ web-registry/src/components/organisms/MediaForm/MediaFormLegacy.tsx
@@ -16,7 +16,8 @@ import {
   ImageUploadProps,
 } from 'web-components/lib/components/inputs/ImageUpload';
 import { VideoInput } from 'web-components/lib/components/inputs/VideoInput';
-import { UrlType } from 'web-components/lib/utils/schemaURL';
+
+import { UrlType } from 'lib/rdf/types';
 
 import getApiUri from '../../../lib/apiUri';
 import { cropAspect, MediaBaseErrors, MediaBaseValues } from './MediaForm';
