+++ web-registry/src/components/organisms/ProjectTopSection/ProjectTopSection.tsx
@@ -22,8 +22,8 @@ import Section from 'web-components/lib/components/section';
 import { TablePaginationParams } from 'web-components/lib/components/table/ActionsTable';
 import { Body, Label, Title } from 'web-components/lib/components/typography';
 
-import { ProjectMetadataIntersectionLD } from 'generated/json-ld';
 import { UseStateSetter } from 'types/react/use-state';
+import { ProjectMetadataIntersectionLD } from 'lib/db/types/json-ld';
 import { getAreaUnit, qudtUnit } from 'lib/rdf';
 
 import { findSanityCreditClass } from 'components/templates/ProjectDetails/ProjectDetails.utils';
