+++ web-registry/src/components/templates/ProjectDetails/ProjectDetails.tsx
@@ -10,6 +10,7 @@ import SEO from 'web-components/lib/components/seo';
 import ProjectMedia from 'web-components/lib/components/sliders/ProjectMedia';
 
 import { useAllCreditClassQuery } from 'generated/sanity-graphql';
+import { ProjectMetadataLD } from 'lib/db/types/json-ld';
 import { getBatchesTotal } from 'lib/ecocredit/api';
 
 import { BuySellOrderFlow } from 'features/marketplace/BuySellOrderFlow/BuySellOrderFlow';
@@ -24,7 +25,6 @@ import {
   useProjectByHandleQuery,
   useProjectByOnChainIdQuery,
 } from '../../../generated/graphql';
-import { ProjectMetadataLD } from '../../../generated/json-ld/index';
 import useEcocreditQuery from '../../../hooks/useEcocreditQuery';
 import useQueryMetadataGraph from '../../../hooks/useQueryMetadataGraph';
 import { useLedger } from '../../../ledger';
