+++ web-registry/src/lib/db/types/json-ld/C02-city-forest-credits-batch.ts
@@ -1,4 +1,4 @@
-import { UrlType } from 'src/utils/schemaURL';
+import { UrlType } from 'lib/rdf/types';
 
 import { BatchMetadataLD } from './batch-base';
