+++ web-registry/src/lib/db/types/json-ld/batch-base.ts
