+++ web-registry/src/lib/db/types/json-ld/credit-batch-intersection-ld.ts
