+++ web-registry/src/lib/rdf/index.ts
@@ -1,2 +1,3 @@
 export * from './rdf';
 export * from './rdf.compacted-context';
+export * from './utils';
