+++ web-registry/src/lib/rdf/types/index.ts
@@ -16,3 +16,12 @@ export interface NameUrl {
 export interface URLList {
   '@list': URL[];
 }
+
+export interface UrlType {
+  '@type': 'schema:URL';
+  '@value'?: string | null;
+}
+
+export interface UrlList {
+  '@list': Array<UrlType>;
+}
