+++ web-registry/src/pages/ProjectMetadata/ProjectMetadata.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { omit } from 'lodash';
 
+import { ProjectMetadataLD } from 'lib/db/types/json-ld';
 import { getProjectShapeIri } from 'lib/rdf';
 
 import {
@@ -13,7 +14,6 @@ import {
   useShaclGraphByUriQuery,
   useUpdateProjectByIdMutation,
 } from '../../generated/graphql';
-import { ProjectMetadataLD } from '../../generated/json-ld';
 import { isVCSCreditClass } from '../../lib/ecocredit/api';
 import { useProjectEditContext } from '../ProjectEdit';
 import { useProjectMetadataSave } from './hooks/useProjectMetadataSave';
