+++ web-registry/src/pages/ProjectReview/hooks/useProjectCreateSubmit.tsx
@@ -1,7 +1,8 @@
 import { useCallback } from 'react';
 import { MsgCreateProject } from '@regen-network/api/lib/generated/regen/ecocredit/v1/tx';
 
-import { ProjectMetadataIntersectionLD } from '../../../generated/json-ld';
+import { ProjectMetadataIntersectionLD } from 'lib/db/types/json-ld';
+
 import { SignAndBroadcastType } from '../../../hooks/useMsgClient';
 import {
   generateIri,
